requirements.txt
appdirs==1.4.3
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
distlib==0.3.0
Django==3.0.7
django-braces==1.14.0
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
drf-multiple-serializer==0.2.3
filelock==3.0.12
idna==2.10
importlib-metadata==1.6.0
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.20
zipp==3.1.0
